add services for products and categories

+++ src/database/models/user.py
@@ -9,6 +9,6 @@ from src.database.session import Base
 class UserModel(Base):
     __tablename__ = 'users'
 
-    user_id: Mapped[BigInteger] = mapped_column(Integer, primary_key=True)
+    user_id: Mapped[BigInteger] = mapped_column(BigInteger, primary_key=True)
     first_name: Mapped[Optional[str]] = mapped_column(String(50), nullable=False)
     description: Mapped[Optional[str]] = mapped_column(String(100), nullable=True, default="Описание вашего профиля")
